+++ Employees/Controllers/SalesController.cs
@@ -11,15 +11,15 @@ namespace Employees.Controllers
         [HttpPost("set-manager")]
         public Manager SetManager([FromBody] Manager manager)
         {
-            Sales sales = new Sales(0.1, 5000, 6500, "Brain", "B", 500);
+            Sales sales = new Sales(0.1m, 5000, 6500, "Brain", "B", 500);
             sales.Manager = manager;
             return sales.Manager;
         }
 
         [HttpGet("calculate-salary")]
-        public double CalculateSalary()
+        public decimal CalculateSalary()
         {
-            Sales sales = new Sales(0.1, 5000, 6500, "Brain", "B", 500);
+            Sales sales = new Sales(0.1m, 5000, 6500, "Brain", "B", 500);
             return sales.CalculateSalary();
         }
     }
